Add Voucher Special apply discount method

+++ src/main/java/rules/PricingRule.java
@@ -5,7 +5,7 @@ import model.Product;
 
 public interface PricingRule {
 
-    long eligibleItems(List<Product> products);
+    Integer eligibleItems(List<Product> products);
 
     Integer apply(List<Product> products);
 }
